Generics, a size() method, and an insertStart() method added.
Any object can be inserted into the Chainlink due to the new implementation of generics into the Chainlink class. The size() method returns the size of a chainlink method from a selected chainlink to the end of the chainlink linked list. The insertStart() method inserts a chain at the start of the chainlink linked list rather than the end.